set ens name
